target/riscv: Make number of debug triggers configurable
In current implementation, the number of debug triggers was hardcoded as
RV_MAX_TRIGGERS macro. This commit replaces the fixed value with a new
"num_triggers" configuration, allowing platforms to configure the number
of debug triggers they want.

If no specific option is provided, the default number of debug triggers
remains 2 as before.

A new CPU property "num-triggers" is added to let users configure the
number of debug triggers from the command line. For example:

  -cpu max,num-triggers=8

Signed-off-by: Alvin Chang <alvinga@andestech.com>
Reviewed-by: Yu-Ming Chang <yumin686@andestech.com>